Move PlannedTransform to dependency-management
